chore: move collection types to types.ts
